Add search view router
Right now we don't have a standardized router. For now, I followed a pattern similar to the router in CodeFileRouter.js which uses the a combination of `history.pushState` and a `popstate` listener. With this, I emulated a pjax-like loading mechanism where the SearchResults component can be injected into the main body of a repo page when the user submits a search.