Fix conflicts between CodeFileRouter and SearchRouter logic
CodeFIleRotuer is a monolithic service, meaning it previously expected to be the only router that inspected and modified the windows historyState. This caused some issues with the new logic in SearchRouter, so some extra logic was added to each router to handle this. This approach of adding monolithic routers is fragile and will not scale, so a centralized router will be needed in the future, but for now this is the fastest way to accomplish what we need.