appdash: Remove panic when missing url in context
This was causing many warnings in the production logs, since the panic was
happening in the http error reporter. This can happen if we return an error
before the appdash middleware has had a chance to run.

All callsites except one already assumed URL could be `nil`, updated the new
callsite.