vfs: Concurrent ReadFile
The VFS interface does not have a batch read interface, so sometimes we are left
with just doing serial reads in our codebase. This CS provides a utility
functions which will add some parallelism to serial reads. Its main feature is
limited read ahead, which allows users of the API to not read everything if they
will break or return early.

Tested with: `godep go test -race -v ./util/vfsutil`. This does not validate
that cleanup occurred unfortunetly. So manually adding accounting and logging
was added to validate that during test runs.