notif: Send emails for discussion actions
We were not sending emails for discussions since it used a fake Person object,
which would not have the Email field set. This commit introduces convenience
functions `notif.Person` and `notif.PersonFromContext` which try to populate as
much useful information about a User for the use in notifications.

`notif.Person` is based on `local.getUserDisplayName`, and will supercede all
uses of it in future.

Test Plan: Run with MANDRILL_KEY set, then comment on a discussion.