app/script/apps/changes/components: when fetching a def has failed, do not cause an exception
When fetching a def fails for any reason `this.state.File` can be invalid or incomplete (missing
it's `RepoRev` field), so catch this early on and do not cause an exception which breaks
CodeReviewPopups in the future.

This doesn't fix the actual issue of why they are not being fetched properly, but it does ensure
that when fetching one fails -- it doesn't break the display of _different ones_ afterwards.