server/local: have RepoTree.Get validate the rev/branch prior to generating tokenized source
This change has the local RepoTree.Get implementation validate that the rev/branch still exists
prior to generating tokenized source code. Prior to this change the tokenized source would have
URLs specific to the rev/branch, which would cause fetching definitions to fail when e.g. viewing
an old diff whose branches have been deleted entirely (although their commits are still in VCS
history). In this event, we now fallback to generating URLs based on the CommitID instead of the
Rev so that fetching these defs does not fail.