util/handlerutil: have GetDefCommon use the Inexact option when user specifies no revision
If the user specifies no Rev when attempting to fetch a definition's common data, then they
do not care about exactness. In the case of fetching definitions for federated code, no Rev
is specified. Thus this makes it such that federated code can be fetched from a build that
is a few commits behind.