server/local: use more concurrency in Deltas.ListFiles
This change addresses a TODO which I've identified as the possible cause of a
perf issue with `Deltas.ListFiles` we're seeing. This is an educated-guess change,
as I cannot reproduce the perf issue locally (potentially because I have an SSD).

My educated guess was made by following the code from the top-level function
(ListFiles) down to possible code paths leading to a Repos.Get or Repos.GetCommit
call (which we're seeing many of). Appdash showed the perf issue to be several
of these Repos.Get* calls in a dependant/consecutive (i.e. non-parralel) order.

From looking at the call graph, this is also a hot spot:

Deltas.ListFiles ->
  Deltas.diff ->
    Deltas.Get ->
      Repos.Get(base)
      Repos.GetCommit(base)
      Repos.Get(head)
      Repos.GetCommit(head)
  parseMultiFileDiffs ->
    each diff ->
      each hunk ->
        linkBaseAndHead ->
          fetchCodeSnippet ->
          fetchCodeSnippet ->
            RepoTree.Get
  formatFileDiffs ->
    each diff ->
      each hunk ->
        formatFileDiffHunk ->
          each chunkDiffOps ->
            RepoTree.Get

Note that formatFileDiffs is not a hot spot like `parseMultiFileDiffs` is Because
it is using concurrency at the `formatFileDiffHunk` level.

Overall I think it is a positive change, or at least not a negative one, and
believe it may fix the perf issue we're seeing when rendering changesets.