events: refactor event definitions and listeners
The reasoning behing this package structure is as follows:

1. Event definitions should be tied to event publisher code
2. Event listeners should be tied to event subscriber code
3. Events should be published from gRPC server endpoints
4. events/ is the server-specific events package. App-specific
event definitions should be placed as a subpackage under the
app's package.

Currently, the git_transport, discussions, and changesets gRPC
server implementations are all within the server/local package.
Hence these event definitions should be within the server-specific
event package, which is events.

Eventually, we want the app-specific gRPC service implementations
to go within the app's package. When that happens for changesets
and discussions, their event definitions should be moved into the
platform app's package too.

Event listeners should be close to the subscriber code. For changesets,
this is the platform app package, so changeset hooks are in there.

When discussions is made into an app, its listeners should move to
its app.

Finally, events/listeners is the server-specific listeners package,
so it will contain git_hooks and any other server-side hooks.