Have canonical URL for apps root page.
Previously, platform apps did not have a canonical URL for the root page,
and the apps themselves could not control this (via redirection) because
the app handler would always see "/" as the request URL.Path.

For example, one could visit issues app via one of:

-	`https://src.example.com/user/repo/.issues`
-	`https://src.example.com/user/repo/.issues/`

The issues app handler would see request URL.Path value of "/" in both cases.

This commit makes it so the canonical URL for app root is the form without
a trailing slash, and enforces it with redirection.

The apps are still oblivious of this and always see "/" as the request
URL.Path value. This is because most routers and users of net/http expect
request URL.Path to always be 1 character or longer and don't support
empty values well.