changesets: don't send notifications for ChangesetUpdateEvent
A ChangesetUpdatEvent is emitted in many cases and not all need to
be notified. For example, if someone pushes to master, all Changesets
with master as the base branch (i.e. mostly every CS), will get
updated. We don't want to populate Slack/email with such notifications.

ChangesetMergeEvent and ChangesetCloseEvent are separate and they
will lead to notifications.