http: Clear context for non-mux routes
This fixes a memory leak. If a route does not go through a mux handler, then it
will never have `gcontext.Clear` called on it. However, we do set a base context
in the outermost middleware. This means that for redirects and static files, we
never clear the context. Note that `gcontext.Clear` is idempotent, so is safe to
call again in the case it has already been called.