changesets: specify author and committer for changeset merges
This fixes and issue where merges would not work on a server without a .gitconfig. This change gathers author information from the changeset.Author field, and committer data from the current user, and uses the RepoStage.Commit method to provide this information to git. The commit/push functionality in RepoStage.Merge function has been removed since it was duplicated in RepoStage.Commit, and the Merge method has been renamed "Pull" since that is more accurate to its function.