server: fix persistence of closed changesets
Instead of attempting to resolve the branch name as soon as the branch is
deleted (which is impossible, because the branch has been deleted) -- we
simply store the proper base revision as soon as the changeset is created.

This has one drawback, which is that if the base is updated (e.g. a entirely
different branch is force-pushed to the same branch name) the change in BASE
will not be detected, and the persisted changeset upon close will reflect
the diff against the old branch / BASE. A simple fix is for users just to
recreate the CS in this case, but we should also be able to detect this
automatically in the future and update the BASE commit ID that we store.