server/internal/store/fs: make Repos.Get check that the directory is actually is repo
Prior to this change we would just check that a Repos.Get operation was for "a valid
directory within ~/.sourcegraph/repos" but not strictly that it was the repository
root. For example, if a repository is symlinked into the repos directory then a
Repos.Get operation would succeed for a subdirectory within the repo (`repo/subdir`)
even though that isn't a repository, but rather a directory inside of it.

While more noticable when the repo is a symlink (i.e. not a bare git repo); it is still
an issue outside of this case. For example visiting:

  https://src.sourcegraph.com/sourcegraph/refs

Produces a "nothing here yet" page because it's not a valid repository yet Repos.Get
responds as if it was (`refs` is a directory inside a bare git repo). We already have
code to detect if something is actually a git repository root, so use that instead
and save ourselves from all of this.