server/internal/store/fs: Fix ReadDir.
Despite the API being stateless, the internal method behavior was
maintaining state. After reading dir with the first call, the dir was
potentially kept open so that future ReadDir calls would return 0
results.

Fix that by seeking to the beginning of directory before reading its
entries.

Use Readdir(0) instead of Readdir(-1) as a style change for
simplicity/consistency. It's a no-op.

Reviewed by @slimsag.