static: Serve assets without auth
To accomplish this we moved the location of the static asset handlers to a
middleware, and put the middleware before the auth middleware.

To test on both `serve-dev` and a release build I browsed the site a bit, and
run curl on a few choice endpoints and ensured I got the expected output.

```
$ curl http://localhost:3080/assets/bundle.j # 404
$ curl http://localhost:3080/assets/bundle.js # 200
$ curl http://localhost:3080/versioned-assets/0.1/bundle.css # 303
$ curl http://localhost:3080/versioned-assets/dev/bundle.css # 200
$ curl http://localhost:3080/favicon.ico # 303
$ curl http://localhost:3080/robots.txt # 200
```