handlerutil: Change util function AugmentCommit to take multiple commits
There are a few places in the codebase where we in a for-loop serially call
AugmentCommit. In a lot of these cases, we repeatedly lookup the same person.
This changes the interface of AugmentCommit to take an array of commits (hence
the rename to AugmentCommits). It then inspects the commits to find the unique
email lookups to do + does them in parallel.

In testing on a large changeset, this sped up a sequential calls to
AugmentCommits from a consistent 80ms, to 20ms uncached, to 4ms cached. (ie 20x
speedup).