build: Ensure private repos are on the GOPATH
This is a workaround for the fact that when we run `go get` in srclib it doesn't
have credentials for fetching private repos. The best solution would involve a
proxy, but instead this solution is easier to get working. We create a GOPATH
per worker which has master checked out for each repo that requires auth.

This feature is disabled on the mothership and for instances which have more
than 50 repos.