platform/apps/changesets: add tests for CLI/UI-based Merge flow
This change adds two new tests:

  TestChangesets_CLIMergeFlow
  TestChangesets_MirroredCLIMergeFlow

which are just a flag to the preexisting MergeFlow tests. When specified
it uses the gRPC Changesets.Merge method instead of using a git merge + push
flow.

This test is currently failing in two cases:

1. The merged status of the CS is not set to merged.
2. The `DeltaSpec.Base.CommitID` is an empty string, i.e. changeset history is not retained.

The tests are valid and correct, as these are legitimate bugs in the `Changeset.Merge`
method. This will be fixed in a followup CS.