idkey: Prevent multi-process race condition on generation
If multiple sourcegraphs are started at the same time backed by the same pgsql
store, then if they need to generate a key then more than one may decide to
generate the key. If this where to happen, only the first instance to write the
key will succeed. The rest will error out and exit. I assume in the case of
multiple instances, there is a load balancer/health system to restart the
instances.