server/internal/middleware/inner: remove support for auto-pulling GitHub profiles/repos
This change removes support for automatically pulling information from GitHub profiles
and repositories. I.e. visiting a user profile that does not exist on the Sourcegraph
instance (but does on GitHub) no longer works. Visiting a user profile no longer shows
the repos from the GitHub user of the same name.

This is a temporary removal until we provide proper integration, allowing a user to specify
their actual GitHub username instead of assuming it's identical to their Sourcegraph one.