app: remove automatic building of unbuilt repos
This feature caused a lot of pain for little gain. It was more useful before we
auto-triggered builds on push and mirror updates. It often results in a lot of CPU
activity and failed builds, which result in a poor user experience.

Also, as we move toward best-effort srclib coverage of commits, the value of this
feature will be further reduced.