server/internal/store/fs: make RepoStage a private changesets-only API
This change marks RepoStage as a changesets-only API, because that is the only user
of it currently. In the future, RepoStage will be entirely done away with and replaced
with just go-vcs with proper repo writing functionality. Until then, adopt this
solution to prevent others from using this RepoStage API.