app/script/components/ExamplesView: match URL pattern used in sourcegraph/def/ExampleView.js
The old ExamplesView.js relies on the old URL pattern (which used a `#`) and thus jumping to
a definition's line range (e.g. clicking `php.go:277-282`) did not work.

This is an intermediate solution until the Changesets frontend code is migrated to use the
new JS architecture, and we can remove the old ExamplesView entirely.