app: do not attempt to update non-mirror repositories
Prior to this change the RefreshVCS operation would be logged verbosely as it
would fail to update the non-mirror repositories inside your SSGPATH/repos (e.g.
hosted or symlinked repos). Change this behavior to avoid the warnings, and
implement a more correct behavior of only enqueuing mirror repos for updates.