platform/apps/changesets: fix race in Merge and Rebase flow tests
Prior to this change the comparison of `ev.Before == cs` was invalid because
`cs` came from a Get operation directly after the merge, and thus hinged on
the fact that the server hadn't yet processed the merge (i.e. the CS was still
marked as unmerged). If the server is going faster (as it does now), this check
fails.