{app,util/handlerutil,ui}: handle empty revs in UI, and use better log messages in backend
This will let us diagnose instances of RepoRevSpecs being passed to
the backend (specifically package sourcecode) better, since it now
logs the repo it was for. (I was unable to reproduce the exact error
@beyang saw.)

This also fixes the UI's handling of empty revs. The expected behavior
is for an empty rev to refer to the repo's default branch. However,
JavaScript was turning an empty rev into the string "undefined" and
requesting a branch named "undefined" from the server, which is
clearly erroneous.