changesets: when merging a hosted repository, go through the git server not FS
Prior to this change we would directly clone from the filesystem and perform
the merge / push changes via the filesystem. This skips the gitserver and thus
githooks do not have a chance to run, and changesets become 'corrupted' as they
do not properly have their merged statuses updated, nor their DeltaSpec updated
(which causes the CS to no longer render after the branch is deleted).