changesets: Fix the broken MergeFlow test implementation
When merging a changeset the valid base commit is _not_ the HEAD after
we commit more changes in the background to master, instead the base commit
should be the point at which we started out branch (i.e. we're not doing a
rebase).

This was previously "passing" but not really working because of the DeltaSpec.Foo != ""
checks which were to be more lenient in the implementation details.

Fixing this test does indeed turn up a valid issue where the Base commit of
a merged CS (via git CLI, src CLI, or via web UI) would not retain history.