app: serve assets via separate handler (not app handler) to avoid nosurf's Vary:Cookie
The nosurf package, used for CSRF protection in the app package, adds
a 'Vary: Cookie' header to all responses. This is OK for app handlers
but not OK for asset handlers because it means they are not cached
whenever a cookie changes. Google Analytics causes the cookies to
change with virtually every request, thus nothing gets cached when
Google Analytics is enabled.

By serving assets through a separate top-level handler, we bypass
nosurf for asset requests.