ui: fix a regression that broke expansion of hunks in changesets
The changesets frontend code makes implicit use of the URL query parameters
to set RepoTreeGetOptions.TreeEntry.FileRange.StartLine/EndLine fields and
without being able to do so hunk expansion is severely broken in changesets.

This regression was caused by 112c30ecf4bd2d676dd9c1cff995fa2f30eddebe

Instead of having this be a general-purpose option (as was the case in 112c30),
make it specific to the UI repo tree endpoint instead.

Setting the Content-Encoding header is also needed or else the content ends up
in the frontend as a string -- not a JS object -- and thus breaks accessing
all fields as it is unparsed.