server/local: properly kick off a build when creating a mirror repo
Prior to this change we would simply clone a mirror repo outright (pgsql
store only) which caused it to have all up-to-date changes in the remote
repo. Thus, when RefreshVCS came around in an attempt to update the repo,
there are no new commits and so no builds are kicked off.

In the case of the fs local store, it initializes an empty git repo and
configures the remote first. This causes a build to fire because RefreshVCS
is responsible for cloning the repo and hence fires off it's githooks,
thus invoking the `buildHook` in `events/git_hooks.go`.

The `TestRepos_CreateStartsBuild_lg` test is now passing.