app: handle platform app redirects properly
Prior to this change if a platform app redirected the user to a different
page (without setting the passthrough header), then it would be considered
"not StatusOK" and thus would be rendered as an error page. This is quite
confusing behavior for users and obviously isn't correct, so instead just
make HTTP redirect status codes equivilent to setting the passthrough
header.