Ensure server is reachable before starting worker
Currently we have a race condition where CI jobs are started before the server
is actually listening properly at the external AppURL address. This leads to
clones failing => build failure. This has been pretty common on k8s for us,
where `kube-proxy` takes a second or two to start proxying the connections. But
in that time we manage to blast through and fail several queued up builds.