changesets: fix wording when there is no user
If an anonymous or unidentified user performed an action, the event
would read: "A user Changeset merged." This was due to a bug in the
handling of message formatting/creation/inflection. This change
simplifies it so that if the user is anonymous, it will simply read "A
user ..." instead of having 2 different sentences, one for identified
users and one for unidentified users.