worker/plan: when there's no srclib data, don't import, just exit
The ideal behavior here would be for the import to occur if the
indexing steps succeeded. Currently we have no way of saying "this
step only runs if these previous steps succeeded."

(The reason why we'd want that behavior is so that if the code was
changed such that none of it was no longer indexed, that fact should
be imported--that none of the code at this version has any refs or
defs. Currently, it is impossible to entirely wipe out code
intelligence data (you have to have some .json files or else the
previous srclib data version's data will remain active because the
import won't occur).