builds: Port Builds_DequeueNext_noRaceCondition to testsuite
Also fixed bug in fs store `Builds.DequeueNext`. It only used a read-lock, but
did mutations. This lead to concurrent calls to `DequeueNext` returning the same
data. The fix still uses a readlock for checking queue length (since empty queue
is the more usual case), but then falls back to a RW lock if popping an item.