issue 76
Merge changeset #614

issue: https://src.sourcegraph.com/sourcegraph/.tracker/76

Added an extra function to track cursor movement as we cannot get cursor position without an event. The issue was that defTooltip does not get the position states until the mouse moves (not just enters and highlights the word). Now we set the initial position in the constructor, which solves the issue.

@Renfredxh