Disable changeset creation button for users without sufficient permissions.
First, fetch the sourcegraph.User for each request and propagate it to
frontend.

This was previously disabled in the past due to performance reasons,
but that was when sg.Users.Get had to go to mothership each time. Now
we have local auth by default, so it should be fast.

The motivation to do this for all pages, rather than just
changesets/compare page, is so a future change can enable
tracing/timing information display for admins on every page.

Add mocks for Users.Get to avoid a nil panic.