changesets: enable filtering of auto-generated files (by default)
Prior to this change large Go changesets that regenerate files (`Godeps/*`,
`*.gen.go`, `*.pb.go`, `*.pb_mock.go`) would be very cumbersome to review and
extremely slow to load (very often the diff is just suppressed entirely). In
order to fix this, we enable adding a `.srcignore` file in this format:

```

Godeps/*
*.gen.go
*.pb.go
*.pb_mock.go
```

Which in turn filters these file diffs out of the Changesets view (identical
frontend behavior to writing a filter in the form of e.g. `/files/!Godeps/` on
the end of a CS URL manually). This file may come from three locations:

1. Repo-specific: You may commit a `.srcignore` file like the above to the
   top-level (root) directory of your repository.
2. Global: If there is no repo-specific `.srcignore`, the global version located
   at `$SGPATH/.srcignore` is used instead.
3. Default: If neither a repo-specific nor global `.srcignore` exists, then a
   default file that works well across most languages is used (right now, this
   is the above example file exactly).