changesets: show which reviewers have given a LGTM with a green checkbox
Prior to this change there was no way to easily see which reviewers on a
changeset had given their LGTM (without looking in the timeline, at least).
After this change a green checkbox appears next to reviewers if they have given
their LGTM (no checkbox if they have not given it). Upon hover, the button still
turns into the red X to remove the reviewer.

---

![img](http://virtivia.com:27080/262bh3om4ezu.png)

---