changesets: have UpdateAffected check for empty struct (instead of nil)
The store `Changesets.Update` operation returns an empty struct (`&sourcegraph.ChangesetEvent{}`)
and not a nil pointer. This makes sense because the exposed gRPC (not store) method
`Changesets.Update` must itself return an empty struct instead of nil (as gRPC does
not allow nil).

Prior to this change, `Changesets.UpdateAffected` would incorrectly check for `nil`
and not an empty struct, and as such these empty structs would leak into the list
returned by `UpdateAffected`.

It's better that we return an empty struct instead of `nil` from the store method,
because there is special (non-store) logic needed for each update and the overall
implementation appears cleaner this way (see follow-up commit after this one).