server/local: fix regression in changesets UpdateAffected operation
This fixes a regression in changesets from when the business logic
was factored out of the store code and moved into `server/local`
instead.

In short, none of this business logic would run during an`UpdateAffected`
operation (which in turn runs several `Update` operations). This was
clearly a regression as full user objects would not be queried (AvatarURLs
not displayed in the frontend, etc).