app: show def tooltip above modal, not behind/obscured by it
Previously the def tooltip would show up behind the def modal, which
made it hard to introspect code using the tooltip in the most
important part of the code (typically what you have the def modal open
for).

This change makes it so the def tooltip shows up ABOVE the modal, so
it is always visible.

Test plan: None. I personally verified it. This change is very minor.