pgsql: include killed builds when dequeueing a new build
When a worker is restarted during a deploy, its active builds are marked
as killed. This commit will ensure that killed builds are considered for
dequeueing when the worker(s) come back online. The build will be started
fresh and all earlier work will be lost.

When dequeueing a killed build, it's ended_at, heartbeat_at, success and
failure fields are reset and would depend on the outcome of the retry.

The tasks from the earlier try of this build are not purged, which means
they will continue showing up on the builds page as 'killed', while new
tasks will be added below them. An improvement will be to delete the old
tasks or hide them from the UI.