changesets: fix viewing a changeset when you are not logged in
Prior to this change it was not possible to view a changeset _that had reviewers_
while you were not logged in. This was because it would attempt to find your status
as a reviewer, which didn't exist, because you are not a reviewer.

After this change, the reviewers panel is still shown to logged-out users, and
clicking the LGTM button has no effect / does not toggle the checkbox.