server/local: prevent list files cache corruption
The list files cache was storing and returning pointers. This can lead
to nasty bugs where two components that get the same hit in the cache
are essentially operating with a global variable.

For example, this snippet of code in the changeset app modified a list
files cache hit:

        if files.OverThreshold {
                for _, fd := range files.FileDiffs {
                        fd.FileDiffHunks = nil
                        fd.FileDiff.Hunks = nil
                }
        }

Now the value is gob encoded on storing in the cache and decoded when
retrieved. This ensures that no modifications on a value being added to
the cache or retrieved from the cache actually affect the values stored
in the cache.

Gob encoding also will allow us to cap the maximum size of value we are
caching. Storing arbitrarly sized values in the cache especially with
regards to diffs could lead to issues that bring down the service.