util/handlerutil: ensure consistent data versions are used when displaying defs
Previously, a bug existed where a def from an older commit C1 could be
shown on a file from a newer commit C2 that has not been
srclib-analyzed. The file would not be srclib-annotated (with refs)
but the def's start/end range would be highlighted. If the file
changed, the def's start/end range could be incorrect (because it
refers to the old commit C1's file).

This commit fixes the issue of inconsistency by being (correctly)
strict about showing defs. If a def at commit C2 (which is not yet
srclib-analyzed) is requested, and if the def's file changed between
C1 and C2, then a 404 is returned. This is the correct but strict
behavior. See the added comments for when this could cause additional
(correct) 404s.

Repro:

1. Create 2 commits, C1 and C2, with a file in a language that srclib
   can analyze. C1 should contain a def. In C2, add lines to the file
   before the def's start/end range.

2. Trigger a build and srclib-analysis for C1. Do not let C2 get
   srclib-analyzed.

3. Visit the URL of the def at C2 (i.e..,
   http://host/REPO@C2/.GoPackage/REPO/MyDef).

Expected: HTTP 404 because C2 is not built and because MyDef's file
changed in C2, so there is no way to show MyDef at commit C2.

Actual: The byte range of MyDef in C1 is highlighted but the file
contents at C2 are used, so the highlighting is misaligned.

Test plan:

Manually verify the steps. The code in util/handlerutil is not easy to
test, but it will be improved in the cheetah changesets.