web_modules/sourcegraph: reset global data after each test run
Prior to this commit, JavaScript frontend tests ran in
nondeterministic order and mutated global state (e.g., when they
dispatched actions like WantDef). This meant that the behavior and
output of tests run subsequently were nondeterministic.

This commit makes it so that global data (in the stores) is reset
before and after each test is executed.

Some autotest output is affected, but that output was erroneous anyway
(it was the result of previous test runs in the same file).

(On my Linux machine `npm test` consistently failed with the same 3
failures. I had ignored these, thinking it was due to some system
configuration issue, but I finally looked into it. I noticed that
`mocha --sort` fixed the issue. That's what made me look into the test
order. It is best, of course, to not rely on deterministic test order
for the tests to be correct; and this commit fixes the underlying
issue.)