repoupdater: Remove enqueueLater
Since we call Enqueue from controllers to ensure a repo has been recently
updated, this leads to a new loop of enqueueLater per page view. ie we have
a constantly restart goroutine per page view of a repo.

Instead we move the responsibility for background updates of mirrored repos to
just `mirrorRepoUpdater`. Note that this implementation is still fragile. We
need to be able to deal with blips speaking to the Repos.List api.