Remove Repo Updater app.
It's currently disabled by default, unless opted-in via
SG_FEATURE_REPOUPDATER. The app is functional and works at "developer"
level of polish, but it's not hitting consumer level of polish at this
time.

Remove the code since it's unused and not a focus at this time; it can
easily be brought back when desired in the future.