server/accesscontrol: Do not short-circuit auth check when hitting public repo.
Previously, verifyPrivateRepoPerms would try to make a decision about
whether the actor had access rights to a given private repo. But in
situations where the repo is public, it doesn't have enough context to
make that decision, so it should defer to the rest of
VerifyActorHas{Read,Write}Access checks. This fix makes that be the
case.

Add a TODO to refactor this to be simpler and safer in the future.

Discussed the auth change design with @pararthshah.