remove repo federation and discovery
To enable us to simplify the implementation of GitHub mirroring and
enable it to be used alongside hosted repos, this commit removes repo
federation and discovery. The notion of "federation root server" still
exists, but it is only used for OAuth2 client registration and
metrics/stats uplink.

The undocumented, confusing Repo.Origin field was also removed, since
it relates to federation.

BREAKING CHANGES: Repo federation no longer works. It will be added
back in another commit before this commit's branch is merged.