util/metricutil: configure metrics forwarding/storing separately from fed config
This change decreases the global significance of the "federation root"
config flag (--fed.root-url), in preparation for removal of that flag.
The concepts of metrics and federation are not inherently related, but
in practice they have been; decoupling them does not change any
behavior.

There is one behavioral change: previously, on the "federation root",
we did not track all events. Now we do (to align the behavior with
that of non-"federation root" servers, since the distinction has been
eliminated for metrics now).