server/local: don't update client name after 1st user signs up
This makes it so that we no longer update the registered client name
on the "federation root" to be the email address of the client
server's first user. This was a convenience for our analytics, but it
had marginal value as learning about the client almost always required
clicking through to see the list of users anyway.

Also, removing it means that nothing in the ./server/local package
behaves in a special way on the "federation root". This commit is one
of a series toward removing the global notion of "federation root"
from the code, to simplify it.