federate repo operations in the app/ui/httpapi/handlerutil layer
Instead of performing repo federation on the backend (in
server/local), which was removed, this commit implements simple repo
federation in the HTTP handler layer.

This is a poor man's federation. It works, it is simple, and it leaves
server/local untouched and totally unaware of federation or even other
servers altogether (which is a good thing).

To test, start a dev server and access
http://localhost:3080/github.com/gorilla/mux (assuming that repo
doesn't exist locally). You should see the repo pulled from
Sourcegraph.com.