pgsql/repos: keep dbRepo.Origin field for automated deployment
This field is kept in the struct to allow automated deployment of
this change. Once the change is deployed, the Origin column will be
removed from the db repo table, and a subsequent commit will revert
this change.

I've tested locally that having the field in the TableMap and deleting
that column from the db is ok, but the other way round (deleting the field
from the TableMap but keeping the db column) is not.