server/internal/middleware/outer: Add a global panic recover to service implementations.
This is meant to be parallel to the global panic handler inside
net/http server for each handler. It doesn't catch all panics (inside
coroutines), but it's a very low cost feature to add. It will hopefully
prevent unexpected panics inside service implementations and libraries.
That way, instead of taking down the entire server, they appear as an
error page.

Include panic stack trace in returned error message. This should be
helpful for determining where the panic came from.