app: simplify def-click behavior
We don't want to do jump-to-def on click right now. This complicated
things with the def card since on click, we need to make sure the card
shows up in the right place (next to the def) and we show the right
file instead of automatically showing the file that corresponds to the
current def. By storing the current "active" def in the store, we don't
have to keep track of this information in the URL, but still retain the
abillity to go-to-definition via a def URL.