circle: force reinstall of git related packages
This seems to be necessary to ensure the latest version of git is
installed. Somehow the preinstalled version of git inside the container
is messing with the package management.

My guess is that when creating the base container Circle CI now first
updates the git version through apt-get then overwrites the bin in some
other operation to the older version of 1.8.5.6. So apt-cache ends up
thinking and reporting via `apt-cache policy` that the latest version
(2.7.2) is already installed skipping the install.

Unskipped the git server tests which were failing because they require
git version 2.3+, because they use the `GIT_SSH` environment variable
to specify the RSA key to use while performing git commands.