pprof: simple routine to dump profiles to disk
Summary:
This is to help debug goroutine leaks in particular. When `SG_SNAPSHOTPROF_PATH`
is set, every 5 minutes a dump of the different profilers is written to the path
in the environment variable. We do not dump the heap profile since it is large
and not as actionable for snapshots.

Test Plan:
Ran

  $ SG_SNAPSHOTPROF_PATH="foo.json" GORACE="" make serve-dev

Average duration for a snapshot on my macbook air was about 0.1s

Reviewers: richard, jonpoler

Reviewed By: jonpoler

Differential Revision: https://srcgraph.phacility.com/D10