repoupdater: enable executing RefreshVCS as a specific user
To refresh private mirror repos, we need to ping the GitHub API as an
authenticated user who has access to that private repo. This change
will allow passing in the current user to repoupdater, which will fire
MirrorRepos.RefreshVCS with that user's token.

Currently, background update of private mirror repos is not supported,
as we do not store which users have access to which repos. This should
be enabled in the future when we do store permissions/ownership
information locally.