worker/plan: Remove unneeded step.
Doing `go get` already fetches, builds and installs the packages. Doing
a `go build` is a subset of `go get`, therefore it has no effect, and
it can be safely removed.

The only other valid reason to keep `go build` step is if the two
commands are executed in sequence, with the 2nd one being executed even
after first one fails. That way, the `go get` step can fail to fetch
dependencies for tests, for example, but `go build` step could succeed
to builld the packages. However, that doesn't appear to be the case -
if the first command fails, the second does not run. That's why I'm
removing it.