app: Redirect home from forgot password when logged in.
Fixes https://app.asana.com/0/30847220579580/99089497784122.

We currently don't support the ability to reset password when already
logged in. Previously, the /forgot endpoint showed a "sorry" message.

However, since it's not supported, there's no normal user flow that can
lead to this page.

It can only happen via redirect-to where if the user starts on a forgot
password page, then clicks "sign up", joins and gets redirected back to
forgot password page. That displays a sorry message, which is not
applicable.

Redirect to home page to prevent that from happening.

Add test to ensure this doesn't regress, since there's no longer a
message rendered for the endpoint when user is logged in.

Test Plan: Tested locally. Added regression test.

Reviewers: beyang

Differential Revision: https://srcgraph.phacility.com/D31