server/local: short circuit root path lookback only when exact commit is specified
Summary:
When a specific commit id is not specified, the GraphStore.Version()
call to get an exact match will fail. So we do need to perform the
lookback starting from the HEAD of the specified branch, in order to
get a valid commit, even for the root path of a repo. This fixes the
issue where package names were not getting resolved.

Test Plan: tested locally, fixes the issue https://app.asana.com/0/87040567695724/98211492240703

Reviewers: beyang

Differential Revision: https://srcgraph.phacility.com/D36