app: use augmented xhr with appdash, auth, cache headers
Summary:
Importing "xhr" gets you a plain-vanilla XHR object without our
special appdash, auth, and cache headers set. This will lead to
unauthenticated requests being made and rejected to
authentication-required endpoints. The only reason we'd want to use
plain "xhr" is for making requests that MUST NOT be authenticated or
MUST NOT use the global cache headers. This is not the case for any of
these places.

This is not an immediate bug for `/.ui/...` endpoints, but it doesn't
hurt, and it is easy to mess up and not provide this auth info when
switching to `/.api/...` endpoints. (API endpoints don't accept cookie
auth, only HTTP basic auth.)

Test Plan: 1. Try some representative pages that fetch data from the affected flux backends.

Reviewers: richard, sqs

Reviewed By: sqs

Subscribers: keegan

Differential Revision: https://srcgraph.phacility.com/D29