app: don't auto-scroll in text results page
Summary:
Fixes an issue where the text search results page would auto-scroll
arbitrarily. This is because it sets the React Blob element's
startLine property, which is necessary for the line numbering but is
also interpreted to mean "the browser should scroll down to the given
line.  That is desirable for the normal code file view but not
elsewhere (examples, search results).

Test Plan:
1. Go to https://sourcegraph.com/github.com/sourcegraph/go-vcs@master/.search?q=VCS&type=text&page=1
2. Notice the browser auto-scrolls down. It should not.

After applying this fix, the browser does not auto-scroll down.

Reviewers: dmitrisg, sqs

Reviewed By: sqs

Differential Revision: https://srcgraph.phacility.com/D41