app: use fetch API instead of XHR
The fetch API can be intercepted by the React bridge, so using fetch
enables us to render React components on the server transparently
(which is fully implemented on the pure-react branch).

Also, fetch is the future standard.

This changeset should be merged in before pure-react to avoid merging
in a ton of stuff at once.

Test plan:

1. Run unit tests
2. Try all pages of the app and look for HTTP errors in the JS console