app: support preloading of React Flux store data for BlobContainer
Adds a new tmpl.Common field "StoreData" that contains preloaded data
for the React Flux stores (only BlobStore and DefStore for now).

This data is added to the store's own internal data upon page load, if
present. This enables immediate rendering of the React components
(without the need for them to fetch data from the server and incur
another roundtrip).