app: pure-Go (no JS) fastpath for rendering blob
Instead of rendering blobs (file contents and annotations such as syntax
highlighting and ref links) in React via JavaScript, we now do the heavy
annotation work in Go. The Blob React component calls to Go on the
server.

This requires that our Go and JavaScript implementations produce EXACTLY
the same HTML. Much of this commit is about achieving that.

The perf win is huge. For a 1,000-line Go file with syntax highlighting
and ref links, the prerender goes from 4sec (which is about what it
takes on the browser as well) to 150msec (and near-instant load in the
browser).