app/coverage: lookback to last built commit for coverage info on repo
Summary:
This will make sure that the coverage dashboard always shows coverage
info for a repo if it has been built previously. Whether Code Intelligence
is currently available on the latest commit of the repo, is a Sourcegraph
specific question and must be dealt with separately.

Also, this change modifies the dashboard to show # of commits that last build
commit is behind the head commit for each repo.

Test Plan: tested locally

Reviewers: beyang

Differential Revision: https://srcgraph.phacility.com/D66