gitserver: Prevent concurrent clones on the same directory
Summary:
Fixes
https://app.asana.com/0/inbox/37478073564517/102011675989608/102088749036651

We likely need more locks than this, but I'd prefer to start off granular (for
performance reasons), and introduce locks as we encounter problems. We have
sufficient monitoring for this strategy.

This diff also introduces a simple package to do locking based on a
`map[string]*sync.Mutex`.

Test Plan: CI + testing clones of larger repos on staging

Reviewers: beyang, richard

Reviewed By: richard

Differential Revision: https://srcgraph.phacility.com/D70